Provide the correct Redis key
Previously, when the user would request to either import or delete the
collections they had created before logging in, the system would try
deleting the collections from the wrong key, leading to the
collections still showing up.

This commit fixes that by providing the code with the correct Redis
key to act upon.